graph_rag/requirements.txt
# Core dependencies for Galaxium Graph RAG System
langchain-core>=0.3.76
langchain-openai>=0.3.33
langchain-graph-retriever>=0.8.0
langchain-astradb>=0.6.1
graph-retriever>=0.8.0
pandas>=2.3.2
python-dotenv>=1.1.1
openai>=1.109.1
tiktoken>=0.11.0

# Optional dependencies for enhanced functionality
jupyter>=1.0.0
matplotlib>=3.7.0
networkx>=3.1.0
plotly>=5.15.0
